==> analysis/step0a_merge_metadata.Rmd <==
---
title: "Merge all metdata and count participants/drawings"
author: "Bria Long"
date: "9/23/2020, finalized 7/2023"
output:
  html_document:
    toc: true
    theme: united
---


# Libraries and setup
```{r}
library(tidyverse)
library(pander)
library(knitr)
library(here)
```

```{r include=FALSE}
knitr::opts_chunk$set(warning=FALSE, message=FALSE)

# Set random seed for reproducibility
set.seed(123)
```

# Count sessions where we had interference

```{r}
interference = read.csv(here::here('data/filtering_outputs/interference_count_across_age_and_runs.csv')) %>%
  mutate(age_numeric = as.numeric(str_split_fixed(interference_age,'age',2)[,2]))
```

```{R}
by_age <- interference %>%
  group_by(age_numeric) %>%
  summarize(parent_draw = sum(parent_draw_count), other_draw = sum(other_draw_count), num_participants = sum(all_surveys_count))
```

# Load and join meta-data for filtered datasets and merged
```{r}
## Load in meta data from filtered cogsci2109 dataset
cogsci_2019 <- read.csv(here::here('data/drawings/cogsci_2019/filtered_dataset_descriptives/cdm_descriptives_nov19_filtered.csv')) %>%
  as.tibble() %>%
  distinct(filename, session_id, category, age, age_numeric, num_strokes, draw_duration_new, draw_duration_old, mean_intensity) %>%
  mutate(filename = str_split_fixed(filename,'/',11)[,11]) %>%
  mutate(filename = str_split_fixed(filename,' ',2)[,2])
```

```{r}
# load and compile metadata from the rest of the runs
which_run = c('cdm_run_v4','cdm_run_v5','cdm_run_v6','cdm_run_v7')
for (run in which_run){
this_run_meta = read.csv(here::here('data/drawings/stringent_cleaned_dataset_meta', paste0(run, '_cleaned.csv'))) %>%
  as.tibble() 
  
  db_records <-read_csv(here::here('data/filtering_outputs/',paste0('images_checked_', run,'.csv'))) %>%
    select(filename, games, numGames)
  
  if (run=='cdm_run_v4'){
    all_db_records <- db_records
    all_meta <- this_run_meta
  }
  else {
  all_db_records <- all_db_records %>%
    full_join(db_records)
  
  all_meta <- all_meta %>%
    full_join(this_run_meta)
  }
  
}
```
# Compute exclusions for paper methods sections
```{r}
cogsci_pre_filter = 15594
cogsci_post_filter = 13205

total_number =  length(unique(all_db_records$filename)) +  cogsci_pre_filter
total_excluded = (cogsci_pre_filter-cogsci_post_filter) + (length(unique(all_db_records$filename))  - length(unique(all_meta$filename)))
total_percent_excluded = total_excluded/total_number
```
Raw drawing data ($N$ = 15594 drawings) were then screened for task compliance using a combination of manual and automated procedures (i.e., excluding blank drawings, pure scribbles, and drawings containing words), resulting in the exclusion of 15.3% of all drawings ($N$ = 13205 drawings after exclusions).

# Write out all meta data, including drawings from cogsci paper
```{r}
all_meta <- all_meta %>%
  full_join(cogsci_2019)

# write_csv(all_meta, path = here::here('data/drawings/stringent_cleaned_dataset_meta', paste0('all_object_metadata_cleaned.csv')))
```

# Appendix A1: Detailed demographics
```{r}
all_meta %>%
  group_by(age_numeric) %>%
  summarize(num_participants = length(unique(session_id)), num_draiwngs = length(unique(filename))) %>%
  kable()
  
```


```{r}
# Option to write out valid drawing sessions (i.e. participant id))

# write_csv(data.frame(valid_sessions), path = here::here('data/drawings/stringent_cleaned_dataset_meta', paste0('valid_sessions.csv')))
```


```{r}
pander(sessionInfo())
```
