==> analysis/step4_production_main_analyses.Rmd <==
---
title: "Drawing production analyses: main analyses, Figure 3/4"
author: "Bria Long"
date: "9/23/2020, finalized 7/2023"
output:
  html_document:
    toc: true
    theme: united
---

# Libraries and setup
```{r setup, include=FALSE}
library(tidyverse)
library(here)
library(assertthat)
library(ggthemes)
# library(egg)
library(viridis)
library(lme4)
library(lmerTest)
# library(MuMIn)
library(langcog)
library(pander)
```

```{r include=FALSE}
knitr::opts_chunk$set(warning=FALSE, message=FALSE)
```

```{r}
# set seed for reproducibility
set.seed(123)
```

# Load main data
## VGG and CLIP classifications
```{r}
load(file=here::here('data/preprocessed_data/merged_vgg_class_and_meta.RData'))
# for each layer
load(file=here::here('data/preprocessed_data/vgg_outputs_by_layer.RData'))

```

```{R}
load(file=here::here('data/preprocessed_data/merged_clip_class_and_meta.RData'))
drawn_categories = unique(d_clip$category)
```

# Drawing recognition increases with age

## Full model predicting accuracy for all drawings
```{R}
all_covariates_accuracy <- glmer(data=d, correct_or_not ~ scale(age_numeric)*scale(drawing_frequency)+scale(avg_tracing_rating) + scale(draw_duration) +
                          scale(mean_intensity) +
                          scale(num_strokes) +
                        (1|session_id) +
                        (1|category), family="binomial",control=glmerControl(optimizer="bobyqa",optCtrl=list(maxfun=2e4)))


```

Plot residuals for main predictor (age)
```{r}
plot(residuals(all_covariates_accuracy) ~ all_covariates_accuracy@frame$`scale(age_numeric)`)
```



Check random effects distribution
```{r}
ranef_plot <- ranef(all_covariates_accuracy, condVar = TRUE)
```

Plot residuals.
```{R}
hist(resid(all_covariates_accuracy, type = "pearson"))
```

```{R}
plot(resid(all_covariates_accuracy)~fitted(all_covariates_accuracy))
```


```{r}
fixed_effect_length = length(summary(all_covariates_accuracy)$coef[,1])
```

```{r}
ci_table1 = confint(all_covariates_accuracy, method='Wald')
```

```{r}
summary_model <- summary(all_covariates_accuracy)

summary_table_complete <- cbind(Estimate = sprintf("%.3f",round(summary_model$coefficients[, "Estimate"],3)),
                                `Std. Error` = sprintf("%.3f",round(summary_model$coefficients[, "Std. Error"],3)),
                                `z value` = sprintf("%.3f",round(summary_model$coefficients[, "z value"],3)),
                                `Pr(>|z|)` = format.pval(summary_model$coefficients[, "Pr(>|z|)"],eps=0.0001,scientific = FALSE, digits=2),
                                 `2.5 % CI` = sprintf("%.3f",round(ci_table1[3:10, 1],3)),
                                `97.5 % CI` = sprintf("%.3f",round(ci_table1[3:10, 2],3)))
                              
```

```{r}

rownames(summary_table_complete) = c('Intercept','Age (in years)','Est. drawing frequency','Avg tracing rating','Time spent drawing','Ink used','Number of strokes','Age*drawing frequency')
```



### Table 1
```{R}
xtable::xtable(as.data.frame(summary_table_complete), digits=4, caption = " \textbf{Modeling children's visual production behavior.}Results of a generalized linear mixed-effect model predicting the recognizability of each drawing (i.e. binary classification scores) from VGG-19 FC6 classifications, including random intercepts for each category and participant. All predictors were z-scored prior to analysis such that coefficients are standardized and comparable. All significance tests are Wald significance tests based on the coefficient values; these tests are two-tailed. No adjustments were made for multiple comparisons. See textit{Methods} for further model specifications.")

```


### Appendix Table: multicolinearity
```{r}
# https://easystats.github.io/blog/posts/performance_check_collinearity/#:~:text=Multicollinearity%20%E2%80%9Cis%20a%20phenomenon%20in,as%20non%2Dindependent%20covariates).
library(performance)
out = check_collinearity(all_covariates_accuracy)
xtable::xtable(out)
```

## Validation: with CLIP
```{r}
all_covariates_accuracy_clip <- glmer(correct_or_not ~ scale(age_numeric)*scale(drawing_frequency)+scale(avg_tracing_rating) + scale(draw_duration) +
                          scale(mean_intensity) +
                          scale(num_strokes) +
                        (1|session_id) +
                        (1|category),
      data = d_clip, family="binomial",control=glmerControl(optimizer="bobyqa",optCtrl=list(maxfun=2e4)))

```

```{r}
cis_clip_production = confint(all_covariates_accuracy_clip, method='Wald')
```

```{r}
summary_model <- summary(all_covariates_accuracy_clip)

summary_table_complete <- cbind(Estimate = sprintf("%.3f",round(summary_model$coefficients[, "Estimate"],3)),
                                `Std. Error` = sprintf("%.3f",round(summary_model$coefficients[, "Std. Error"],3)),
                                `z value` = sprintf("%.3f",round(summary_model$coefficients[, "z value"],3)),
                                `Pr(>|z|)` = format.pval(summary_model$coefficients[, "Pr(>|z|)"],eps=0.0001,scientific = FALSE, digits=2),
                                 `2.5 % CI` = sprintf("%.3f",round(cis_clip_production[3:10, 1],3)),
                                `97.5 % CI` = sprintf("%.3f",round(cis_clip_production[3:10, 2],3)))
                              
```

```{r}

rownames(summary_table_complete) = c('Intercept','Age (in years)','Est. drawing frequency','Avg tracing rating','Time spent drawing','Ink used','Number of strokes','Age*drawing frequency')
```


### Appendix: CLIP classifications
```{R}
xtable::xtable(summary_table_complete, digits=4, caption = "Model coefficients of a GLMM predicting the recognizability of each drawing (i.e. binary classification scores) from CLIP classifications, including random intercepts for each category and participant.")

```

# More frequently drawn categories are not more recognizable. 

## Load, anlayze, and plot category frequency ratings
```{r}
frequency = read_csv(file = here::here('data/surveys/drawing_experience/preprocessed/Category_frequency_survey.csv')) 
```

To assess this, `r length(unique(frequency$subject_id))` parents of children aged 3-10 years filled out a survey asking about the frequency with with their children drew the categories in the dataset.

```{r}
count_by_age <- frequency %>%
  group_by(childs_age) %>%
  dplyr::summarize(num_surveys = length(unique(subject_id)))
```

```{r}
freq_by_category_temp<- frequency %>%
  # filter(childs_age > 2) %>%
  mutate(category = str_split_fixed(category,' ',2)[,2]) %>%
  mutate(category = str_replace(category,' ','.')) %>%
  filter(category %in% drawn_categories) 

freq_by_category <- freq_by_category_temp %>%
  group_by(category) %>%
  summarize(drawing_frequency = mean(often_drawn_rating)) 

# write_csv(freq_by_category, here::here('data/surveys/drawing_experience/preprocessed/freq_by_category.csv'))
```

```{r}
freq_by_category_plot <- freq_by_category_temp %>%
  group_by(category, childs_age) %>%
  multi_boot_standard(col = 'often_drawn_rating') %>% 
  ungroup() %>%
  mutate(category = fct_reorder(category, mean, .desc=TRUE)) 

freq_by_category_by_age <- freq_by_category_temp %>%
  group_by(category, childs_age) %>%
  summarize(frequency = mean(often_drawn_rating))
```

```{R}
# summary(lm(data=freq_by_category_by_age, frequency ~ childs_age*category))
```

```{r}
ggplot(freq_by_category_plot, aes(x=category, y=mean, col=childs_age)) +
  theme_few(base_size=10)+
  geom_jitter(height=.02, width=0, alpha=.2, size=.6)+
  theme(axis.text.x = element_text(angle = 90, vjust = 0.5, hjust=1)) +
  labs(y = 'Estimated drawing frequency \n 1 = Never, 5 = Everyday', x  = '') +
  ylim(.5,5.5) +
  geom_smooth(aes(group=childs_age),alpha=.02, span=10) +
  scale_color_viridis_c(name = 'Child\'s age') +
  theme(legend.position = 'right')

ggsave(here::here('data/figures_out/Appendix_A6_estimated_frequency.pdf'), units='in', height=4, width=8)

write_csv(freq_by_category_plot, file = here::here('data/figures_csvs/Appendix_A6_estimated_frequency.csv'))


```


## Load other measures of frequency
```{r}
aoas <- readRDS(file=here::here('data/category_features/english_(american)_aoa_bydefinition.rds')) %>%
  filter(measure=="produces") %>%
  filter(definition %in% drawn_categories) %>%
  rename(word = definition)
```

```{r}
load(file = here::here('data/category_features/merged_word_freqs.Rdata'))
category_by_word_freq <- ch_freq_smooth %>%
  filter(word %in% drawn_categories) %>%
  filter(lexical_class == 'nouns') %>%
  distinct(word, on_cdi, CHILDES, adult_books)
  
```
### Join all frequnecy data together
```{r}
all_freq_by_category <-  freq_by_category %>%
  left_join(category_by_word_freq, by=c('category' = 'word')) %>%
  left_join(aoas %>% select(word, aoa), by=c('category' = 'word'))
 
```

```{r}
d_with_freq <- d %>%
  left_join(all_freq_by_category)
```

```{r}
all_freq_covariates <-  glmer(correct_or_not ~ scale(age_numeric)+scale(adult_books) + scale(aoa) + scale(CHILDES) + scale(drawing_frequency) +
                        (1|session_id) +
                        (1|category),
      data = d_with_freq , family="binomial",control=glmerControl(optimizer="bobyqa",optCtrl=list(maxfun=2e4)))

cis_freq = confint(all_freq_covariates, method='Wald')



```
```{R}

summary_model = summary(all_freq_covariates)

summary_table_complete_freq <- cbind(Estimate = sprintf("%.3f",round(summary_model$coefficients[, "Estimate"],3)),
                                `Std. Error` = sprintf("%.3f",round(summary_model$coefficients[, "Std. Error"],3)),
                                `z value` = sprintf("%.3f",round(summary_model$coefficients[, "z value"],3)),
                                `Pr(>|z|)` = format.pval(summary_model$coefficients[, "Pr(>|z|)"],eps=0.0001,scientific = FALSE, digits=2),
                                 `2.5 % CI` = sprintf("%.3f",round(cis_freq[3:8, 1],3)),
                                `97.5 % CI` = sprintf("%.3f",round(cis_freq[3:8, 2],3)))

rownames(summary_table_complete_freq)=c('Intercept','Age (in years)','Freq in adult books','Age-of-Acquisition','CHILDES frequency','Drawing frequency')



  
```

## Appendix: Table A5: No effect of other frequency  measures on recognizability 


```{R}
xtable::xtable(summary_table_complete_freq, digits=3, caption = "Model coefficients of a GLMM predicting the recognizability of each drawing (i.e. binary classification scores) from children's age, the frequency of each category in CHILDES (data for 91.8 percent of categories), the estimated Age-of-Aquisition using WordBank data (data for 73 percent of categories), and the frequency of each word in adult English books. All predictors are z-scored; random intercepts for each category and participant are included.")

```

```{r}
# clip_correct_by_category <- d_clip %>%
#   group_by(category) %>%
#   summarize(prop_correct = mean(correct_or_not)) %>%
#   left_join(category_by_word_freq, by=c('category' = 'word')) %>%
#   left_join(aoas) %>%
#   mutate(aoa = replace_na(aoa, 48))
```

```{r}
# frequency in childes
# cor.test(clip_correct_by_category$prop_correct, clip_correct_by_category$CHILDES)

# frequency in adult books (google ngrams)
# cor.test(clip_correct_by_category$prop_correct, clip_correct_by_category$adult_books)

# average age of acquisition (for the 31 categories on CDI)
# cor.test(clip_correct_by_category$prop_correct, clip_correct_by_category$aoa)

```





## Appendix: Table A4: No Effect of drawing frequency in human recognition data
```{r}
### Load human classification data

humans <- read.csv(here::here('data/devphotodraw_recognition_ratings/compiled_human_recognition.csv')) 
  
devphotodraw_recog <- humans %>%
  mutate(condition = str_split_fixed(image_name_short,'_',2)[,1]) %>%
  filter(condition == 'S') %>%
  mutate(age = replace(age,age==10, 9)) %>%
  group_by(unique_ids, category, age, image_name_short) %>% # group by each drawing of each category
  summarize(prop_correct = mean(correct_or_not))  %>%
  left_join(freq_by_category)
```

```{r}
devphotodraw_recog_out = lmer(data = devphotodraw_recog, prop_correct ~ scale(drawing_frequency)*scale(age) + (1|unique_ids) + (1|category))

cis_devphotodraw = confint(devphotodraw_recog_out, method='Wald')

```

```{r}
summary_model = summary(devphotodraw_recog_out)

summary_table_devphotodraw <- cbind(Estimate = sprintf("%.3f",round(summary_model$coefficients[, "Estimate"],3)),
                                `Std. Error` = sprintf("%.3f",round(summary_model$coefficients[, "Std. Error"],3)),
                                                                `df` = sprintf("%.3f",round(summary_model$coefficients[, "df"],3)),
                                `t value` = sprintf("%.3f",round(summary_model$coefficients[, "t value"],3)),
                                `Pr(>|t|)` = format.pval(summary_model$coefficients[, "Pr(>|t|)"],eps=0.001,scientific = FALSE, digits=2),
                                 `2.5 % CI` = sprintf("%.3f",round(cis_devphotodraw[4:7, 1],3)),
                                `97.5 % CI` = sprintf("%.3f",round(cis_devphotodraw[4:7, 2],3)))

rownames(summary_table_devphotodraw)=c('Intercept','Est. drawing frequency','Age (in years)','Age*drawing frequency')
```

```{R}
xtable::xtable(summary_table_devphotodraw, digits=3, caption = "Model coefficients of a linear mixed effect model predicting the recognizability of each drawing (as assesed by crowd-sourced adult behavioral data). Drawings were produced in an experimental context. All predictors are z-scored and random intercepts for each category and participant are included.")

```


# Recognizable drawings contain more category-diagnostic information across development
```{r}
correct_subset <- d  %>% 
  filter(correct_or_not==1)

all_covariates_log_odds_corr_only <- lmer(log_odds ~ scale(age_numeric)*scale(drawing_frequency) + scale(avg_tracing_rating) + scale(draw_duration) +
                          scale(mean_intensity) +
                          scale(num_strokes) +
                        (1|session_id) +
                        (1|category),
      data = correct_subset, control=lmerControl(optimizer="bobyqa",optCtrl=list(maxfun=2e4)))

log_odds_out = summary(all_covariates_log_odds_corr_only)$coef

all_covariates_log_odds_corr_only_ci = confint(all_covariates_log_odds_corr_only, method='Wald')

```

```{r}
correct_subset_clip <- d_clip %>% 
  filter(correct_or_not==1)

all_covariates_log_odds_corr_only_clip <- lmer(log_odds ~ scale(age_numeric)*scale(drawing_frequency) + scale(avg_tracing_rating) + scale(draw_duration) +
                          scale(mean_intensity) +
                          scale(num_strokes) +
                        (1|session_id) +
                        (1|category),
      data = correct_subset_clip, control=lmerControl(optimizer="bobyqa",optCtrl=list(maxfun=2e4)))

log_odds_out_clip = summary(all_covariates_log_odds_corr_only_clip)$coef


```



Even for drawings that were correctly classified (`r round(length(unique(correct_subset$filename))/length(unique(d$filename)),3)` of the balanced subset of drawings, $N$=`r length(unique(correct_subset$filename))`) there was still a reliable increase in the amount of diagnostic information they contained across age, as measured by the log-odds probability assigned by the logistic-regression classifier to the target category (see \textit{Methods}, B = `r round(log_odds_out[2,1],3)` , SE = `r round(log_odds_out[2,2],3)`, df =  `r round(log_odds_out[2,3],3)`, t = `r round(log_odds_out[2,4],3)`, P < `r round(log_odds_out[2,5],3)`, CI = [`r round(all_covariates_log_odds_corr_only_ci[5,1],3)`,`r round(all_covariates_log_odds_corr_only_ci[5,2],3)`]). Thus, this secondary analysis provides additional support for the account that the age-related improvements reflect gradual changes in children's ability to include diagnostic visual features in their drawings. 



### Appendix: Full coefficient tables for correctly classified drawings (VGG)
```{r}

cis_log_odds_vgg = confint(all_covariates_log_odds_corr_only, method='Wald')
summary_model <- summary(all_covariates_log_odds_corr_only)

summary_table_complete_cis_log_odds_vgg <- cbind(Estimate = sprintf("%.3f",round(summary_model$coefficients[, "Estimate"],3)),
                                `Std. Error` = sprintf("%.3f",round(summary_model$coefficients[, "Std. Error"],3)),
                                                                `df` = sprintf("%.3f",round(summary_model$coefficients[, "df"],3)),
                                `t value` = sprintf("%.3f",round(summary_model$coefficients[, "t value"],3)),
                                `Pr(>|t|)` = format.pval(summary_model$coefficients[, "Pr(>|t|)"],eps=0.001,scientific = FALSE, digits=2),
                                 `2.5 % CI` = sprintf("%.3f",round(cis_log_odds_vgg[4:11, 1],3)),
                                `97.5 % CI` = sprintf("%.3f",round(cis_log_odds_vgg[4:11, 2],3)))


rownames(summary_table_complete_cis_log_odds_vgg) = c('Intercept','Age (in years)','Est. drawing frequency','Avg tracing rating','Time spent drawing','Ink used','Number of strokes','Age*drawing frequency')                              
```

```{R}
xtable::xtable(summary_table_complete_cis_log_odds_vgg, digits=3, caption = "Model coefficients of a linear mixed effect model predicting the log-odds probability assigned to correctly classified drawings using VGG embeddings, including random intercepts for each category and participant.")

```

### Appendix: Full coefficient tables for correctly classified drawings (CLIP)

Summary of model predicting clip log-odds for correctly recognized drawings.
```{r}

cis_log_odds_clip = confint(all_covariates_log_odds_corr_only_clip, method='Wald')
summary_model <- summary(all_covariates_log_odds_corr_only_clip)

summary_table_complete_cis_log_odds_clip <- cbind(Estimate = sprintf("%.3f",round(summary_model$coefficients[, "Estimate"],3)),
                                `Std. Error` = sprintf("%.3f",round(summary_model$coefficients[, "Std. Error"],3)),
                                                                `df` = sprintf("%.3f",round(summary_model$coefficients[, "df"],3)),
                                `t value` = sprintf("%.3f",round(summary_model$coefficients[, "t value"],3)),
                                `Pr(>|t|)` = format.pval(summary_model$coefficients[, "Pr(>|t|)"],eps=0.001,scientific = FALSE, digits=2),
                                 `2.5 % CI` = sprintf("%.3f",round(cis_log_odds_clip[4:11, 1],3)),
                                `97.5 % CI` = sprintf("%.3f",round(cis_log_odds_clip[4:11, 2],3)))


rownames(summary_table_complete_cis_log_odds_clip) = c('Intercept','Age (in years)','Est. drawing frequency','Avg tracing rating','Time spent drawing','Ink used','Number of strokes','Age*drawing frequency')                              
```



```{R}
xtable::xtable(summary_table_complete_cis_log_odds_clip, digits=3, caption = "Model coefficients of a linear mixed effect model predicting the log-odds probability assigned to correctly classified drawings using CLIP embeddings, including random intercepts for each category and participant.")

```

#Visuomotor control explains some but not all of changes in drawing recognizability
Model coefficients in Table 1, referneced above.

## Appendix Plot A7:  Drawing recognizability by effort & tracing
How does classification accuracy interact with age x effort & tracing ?
```{r}
num_bins = 3

cor_by_age_by_strokes <- d %>%
  ungroup() %>%
  mutate(bin = ntile(num_strokes, num_bins)) %>%
  group_by(bin, age_numeric,category) %>%
  summarize(avg_cor = mean(correct_or_not)) %>%
  group_by(age_numeric,bin) %>%
  multi_boot_standard(col = "avg_cor")   %>%
  mutate(covariate = 'by strokes drawn')

cor_by_age_by_time <- d %>%
  ungroup() %>%
  mutate(bin = ntile(draw_duration, num_bins)) %>%
  group_by(bin, age_numeric,category) %>%
  summarize(avg_cor = mean(correct_or_not)) %>%
  group_by(age_numeric,bin) %>%
  multi_boot_standard(col = "avg_cor")  %>%
  mutate(covariate = 'by time spent')

cor_by_age_by_intensity <- d %>%
  ungroup() %>%
  mutate(bin = ntile(mean_intensity, num_bins)) %>%
  group_by(bin, age_numeric,category) %>%
  summarize(avg_cor = mean(correct_or_not)) %>%
  group_by(age_numeric,bin) %>%
  multi_boot_standard(col = "avg_cor")   %>%
  mutate(covariate = 'by ink used')

cor_by_age_by_tracing <- d %>%
  ungroup() %>%
  filter(!is.na(avg_tracing_rating)) %>%
  mutate(bin = ntile(avg_tracing_rating, num_bins)) %>%
  group_by(bin, age_numeric,category) %>%
  summarize(avg_cor = mean(correct_or_not)) %>%
  group_by(age_numeric,bin) %>%
  multi_boot_standard(col = "avg_cor")   %>%
  mutate(covariate = 'by tracing ability')
```

Merge data together
```{r}
merged <- cor_by_age_by_intensity %>%
  full_join(cor_by_age_by_strokes) %>%
  full_join(cor_by_age_by_time) %>%
  full_join(cor_by_age_by_tracing) %>%
  mutate(bin_name = as.numeric(bin))
```

Make the plot
```{r}
smooth_alpha=.1
ggplot(merged, aes(age_numeric,mean*100, color=bin_name, group=bin_name, col=bin_name)) +
  geom_pointrange(aes(ymin = ci_lower*100, ymax = ci_upper*100), alpha=.6, size=.25) +
  theme_few(base_size = 10) +
  labs(x='Age of child drawing (yrs)', y='Category classification \n accuracy') +
  scale_color_viridis(option="C", begin=.2, end=.8, discrete=FALSE, name = 'Effort') +
  scale_x_continuous(breaks=seq(2,10,1)) +
  geom_smooth(span=10, alpha=smooth_alpha, size=.4) +
  ylim(0,100) +
  geom_hline(yintercept = 1/48, linetype="dashed", color="grey") +
  theme(legend.position ='none',aspect.ratio = 1) +
  scale_fill_discrete(labels=c('Low','Medium','High')) + 
  facet_grid(~covariate)
  

ggsave(here::here('data/figures_out/Appendix_A7_visuomotor_control_wide.pdf'), units='in')

write_csv(merged, here::here('data/figures_csvs/Appendix_A7_visuomotor_control_wide.csv'))

```



# Unrecognizable drawings still contain semantically relevant information
### Make long form classification data
```{r}
classification_data_long <- all_layer_class %>%
  filter(layer=='FC6') %>%
  mutate(correct_or_not = as.logical(correct_or_not)) %>%
  gather(key = 'class', value = 'prob', contains('prob')) %>%
  mutate(class = str_split_fixed(class, '_prob',2)[,1])
```



### Load category animacy/size (hardcoding)
```{r}
animacy_csv <- read_csv(here::here('data/drawings/category_metadata/animacy.csv')) %>%
  as_tibble() %>%
  mutate(animacy_size = case_when(animacy == '0' & size=='0' ~ 4, # small objects
                                  animacy == '0' & size=='1' ~ 3, # big objects
                                  animacy == '1' & size=='0' ~ 2, # small animals
                                  animacy == '1' & size=='1' ~ 1)) # big animals

## ANIM/SIZE ORDERING IS JUST ABOUT HOW YOU WANT CATEGORIES PLOTTED.
```


### Calculate avg prob by category for misclassified
```{r}
observations_count <- classification_data_long %>%
  filter(correct_or_not==0) %>%
  ungroup() %>%
  summarize(num_drawings = length(unique(image_name)), num_subs = length(unique(session_id)))
```

We analyzed N=`r observations_count$num_drawings` from `r observations_count$num_subs` that were misclassified.


```{r}
observations_count_clip <- d_clip %>%
  filter(correct_or_not==0) %>%
  ungroup() %>%
  summarize(num_drawings = length(unique(filename)), num_subs = length(unique(session_id)))
```


```{r}
confusions_by_class <- classification_data_long %>%
  mutate(drawn_category = category) %>%
  left_join(animacy_csv) %>%
  ungroup() %>%
  filter(correct_or_not==0) %>%
  mutate(drawn_category = fct_reorder(drawn_category, animacy_size)) %>%
  mutate(class = factor(class, levels = levels(drawn_category))) %>%
  group_by(drawn_category, class) %>%
  dplyr::summarize(mean_prob = mean(prob))



```


### Figure 3A: Confusions by class
```{r}
ggplot(data = confusions_by_class, aes(x=class, y=drawn_category,fill=mean_prob)) + 
  geom_tile() + 
  theme_few(base_size = 12) +
  theme(legend.position = 'none', axis.text.x = element_text(angle = 90,  
  size = 6, vjust=.5, hjust = 1), axis.text.y = element_text(angle = 0,
  size = 6, hjust = 1)) + 
  coord_fixed() + 
  scale_fill_viridis(option="A", name = 'Classifier probability')  +
  ylab('Target category') +
  xlab('Predicted category') 

ggsave(here::here('data/figures_out/Figure3A_classifier_confusions.pdf'), width = 7.5, height = 5, units = 'in')

write_csv(confusions_by_class, file = here::here('data/figures_csvs/Figure3A_classifier_confusions.csv'))
```



### Calculate whether animacy/size were correct
```{r}
animacy_size_acc_by_age <- classification_data_long %>%
  mutate(drawn_category = category) %>%
  left_join(animacy_csv %>% select(animacy, size, category), by = c("drawn_category" = "category")) %>%
  ungroup() %>%
  rename(drawn_category_animacy = animacy, drawn_category_size = size) %>%
  left_join(animacy_csv %>% select(animacy, size, category),  by = c("class" = "category")) %>%
  rename(class_animacy = animacy, class_size = size) %>%
  group_by(image_name, drawn_category, age_numeric, layer) %>%
  mutate(max_prob = max(prob), top_class = prob==max_prob)  %>%
  filter(top_class==TRUE) %>%
  mutate(animacy_correct = (drawn_category_animacy == class_animacy)) %>%
  mutate(size_correct = (drawn_category_size == class_size)) 

```

## Examine category variation for animacy classification
```{r}
baseline_animacy <- mean(animacy_csv$animacy)
baseline_objects <- 1 - mean(animacy_csv$animacy)

animacy_acc_sanity <- animacy_size_acc_by_age %>%
  filter(correct_or_not==FALSE) %>%
  group_by(category, drawn_category_animacy) %>%
  summarize(prop_animacy_correct = mean(animacy_correct)) %>%
  mutate(baseline_chance = case_when(drawn_category_animacy == 0 ~ baseline_objects, 
                                     drawn_category_animacy == 1 ~ baseline_animacy)) %>%
  mutate(anim_acc = prop_animacy_correct - baseline_chance) %>%
  arrange(-anim_acc)
```




### Calculate animacy correct relative to basline
```{r}
baseline_animacy <- mean(animacy_csv$animacy)
baseline_objects <- 1 - mean(animacy_csv$animacy)

anim_cor_by_category_by_age <- animacy_size_acc_by_age %>%
  filter(correct_or_not==0) %>%
  mutate(drawn_category_animacy = as.factor(drawn_category_animacy))  %>%
  mutate(baseline_chance = case_when(drawn_category_animacy == 0 ~ baseline_objects, 
                                     drawn_category_animacy == 1 ~ baseline_animacy)) %>%
  group_by(age_numeric, drawn_category, layer) %>%
  dplyr::summarize(avg_animacy_correct = mean(animacy_correct) - baseline_chance) %>%
  distinct(age_numeric, avg_animacy_correct, drawn_category)

  
anim_cor_by_age <- anim_cor_by_category_by_age %>%
  group_by(age_numeric, layer) %>%
  multi_boot_standard(col = 'avg_animacy_correct')

```


### Figure 3B: Plot animacy correct by age
```{r}
smooth_alpha=.2
ggplot(anim_cor_by_age, aes(x=age_numeric,y=mean, color=age_numeric)) +
  geom_pointrange(aes(ymin = ci_lower, ymax = ci_upper), alpha=.8, size=.2) + 
  geom_point(data=anim_cor_by_category_by_age, aes(x=age_numeric, y=avg_animacy_correct), alpha=.1, size=.5) +
  theme_few(base_size = 8) + 
  labs(x='Age', y='Prop. animacy correct') +
  theme(legend.position = "none") +
  # geom_smooth(span=20, alpha=smooth_alpha, color='grey') +
  scale_color_viridis(option="D", discrete=FALSE) +
  geom_hline(yintercept =  0 , linetype="dashed", color="grey")  
  # facet_grid(~layer)


ggsave(here::here('data/figures_out/Figure3b_animacy_classification.pdf'), width = 1.5, height = 1.5, units = 'in')

write_csv(anim_cor_by_age, file = here::here('data/figures_csvs/Figure3b_animacy_classification.csv'))

write_csv(anim_cor_by_category_by_age, file = here::here('data/figures_csvs/Figure3b_animacy_classification_by_category.csv'))


```

```{r}
objects_only <- animacy_csv %>%
  filter(animacy==0)

animals_only <- animacy_csv %>%
  filter(animacy==1)

baseline_big <- mean(objects_only$size)
baseline_small <- 1 - mean(objects_only$size)

baseline_big_animals <- mean(objects_only$size)
baseline_small_animals <- 1 - mean(animals_only$size)

baseline_big_all <- mean(animacy_csv$size)
baseline_small_all <- 1 - mean(animacy_csv$size)
```

```{r}
size_cor_by_category_by_age <- animacy_size_acc_by_age %>%
  filter(correct_or_not==0) %>%
  filter(drawn_category_animacy==0) %>% #only objects
  mutate(drawn_category_size = as.factor(drawn_category_size))  %>%
  mutate(baseline_chance = case_when(drawn_category_size == 0 ~ baseline_small_all, 
                                     drawn_category_size == 1 ~ baseline_big_all)) %>%
  group_by(age_numeric, drawn_category, baseline_chance) %>%
  summarize(avg_size_correct = (mean(size_correct))) %>%
  mutate(avg_size_correct = avg_size_correct - baseline_chance)

size_cor_by_age <- size_cor_by_category_by_age %>%
  group_by(age_numeric) %>%
  multi_boot_standard(col = 'avg_size_correct')
```


```{r}
count_drawings_size <- animacy_size_acc_by_age %>%
  filter(correct_or_not==0) %>%
  filter(drawn_category_animacy==0) %>%
  summarize(num_drawings= length(unique(image_name)))
```


Not true for animal size! (not in paper but interesting)
```{r}
animal_size_cor_by_category_by_age <- animacy_size_acc_by_age %>%
  filter(correct_or_not==0) %>%
  filter(drawn_category_animacy==1) %>% #only animals
  mutate(drawn_category_size = as.factor(drawn_category_size))  %>%
  mutate(baseline_chance = case_when(drawn_category_size == 0 ~ baseline_small_animals, 
                                     drawn_category_size == 1 ~ baseline_big_animals)) %>%
  group_by(age_numeric, drawn_category, baseline_chance) %>%
  summarize(avg_size_correct = (mean(size_correct))) %>%
  mutate(avg_size_correct = avg_size_correct - baseline_chance)

animal_size_cor_by_age <- animal_size_cor_by_category_by_age %>%
  group_by(age_numeric) %>%
  multi_boot_standard(col = 'avg_size_correct')
```

Look at category variation in object size classifications.
```{R}
size_cor_by_category <- animacy_size_acc_by_age %>%
  filter(correct_or_not==0) %>%
  filter(drawn_category_animacy==0) %>% #only objects
  mutate(drawn_category_size = as.factor(drawn_category_size))  %>%
  mutate(baseline_chance = case_when(drawn_category_size == 0 ~ baseline_small_all, 
                                     drawn_category_size == 1 ~ baseline_big_all)) %>%
  group_by(drawn_category, drawn_category_size) %>%
  summarize(avg_size_correct = mean(size_correct) - baseline_chance)  %>%
  distinct(avg_size_correct, drawn_category)
```

### Figure 3C: Size correct by age

```{r}


ggplot(size_cor_by_age, aes(x=age_numeric,y=mean, color=age_numeric)) +
  geom_pointrange(aes(ymin = ci_lower, ymax = ci_upper), alpha=.8, size=.2) + 
  geom_point(data=size_cor_by_category_by_age, aes(x=age_numeric, y=avg_size_correct), alpha=.1, size=.5) +
  theme_few(base_size = 8) + 
  labs(x='Age', y='Prop. object size correct') +
  theme(legend.position = "none") +
  # geom_smooth(span=10, alpha=smooth_alpha, color='grey') +
  scale_color_viridis(option="D", discrete=FALSE) +
  geom_hline(yintercept =  0 , linetype="dashed", color="grey") 
  
ggsave(here::here('data/figures_out/Figure3b_size_classification.pdf'), width = 1.5, height = 1.5, units = 'in')

write_csv(size_cor_by_age, file = here::here('data/figures_csvs/Figure3b_size_classification.csv'))

write_csv(size_cor_by_category_by_age, file = here::here('data/figures_csvs/Figure3b_size_classification_each_category.csv'))
```


## Create t-tests tables if desired
```{r}
library(dplyr)
library(broom)
library(tibble)

animacy_tests <- anim_cor_by_category_by_age %>%
  group_by(age_numeric) %>%
  summarise(t_test_result = list(tidy(t.test(avg_animacy_correct)))) %>%
  unnest(t_test_result) 


size_tests <- size_cor_by_category_by_age %>%
  group_by(age_numeric) %>%
  summarise(t_test_result = list(tidy(t.test(avg_size_correct)))) %>%
  unnest(t_test_result)  %>%
  select(-method, -alternative) 

```


```{R}
# xtable::xtable(as.data.frame(size_tests), digits=4, caption = "Individual one-sample t-tests performed for each age over the animacy and object size classifications values for all 48 categories (aniacy) or the 25 inanimate categories (object size)")

```


# Drawings contain different semantic parts across development
```{r}
# all annotations before filtering
load(file = here::here('data/part_annotations_processed/merged_annotations.RData'))
```

```{r}
# preprocesssed emphasis data
load(file = here::here('data/part_annotations_processed/part_emphasis.RData'))
```

```{r}
# example drawing in presentation, or any other we want to examine
# example <- part_emphasis_all %>% filter(filename == 'bird_sketch_age8_cdm_run_v31530484377182.png')
# 
# what_annotations <- d_agree_labels %>% filter(filename == 'bird_sketch_age8_cdm_run_v31530484377182.png')
```
## Get data structure with unique labels per drawing
```{r}
unique_labels <- d_agree_labels %>%
  group_by(filename,category, age_numeric, gameType) %>%
    summarize(count_parts = length(unique(roi_labelName)), count_unint = sum(sum(roi_labelName == 'unintelligible')>0)) %>%
  # only count unintellible as one label since looking at # of unique labels
  mutate(num_labels = count_parts - count_unint) # 
```


## Plot of frequency of unique labels by category and age
```{r include=TRUE}
unique_by_age_by_category <- unique_labels %>%
  group_by(age_numeric, category) %>%
  multi_boot_standard(col = 'num_labels')

unique_by_age <- unique_labels %>%
  group_by(age_numeric) %>%
  multi_boot_standard(col = 'num_labels')
```


```{r}
ggplot(data=unique_by_age_by_category, aes(x=age_numeric, y=mean, color=age_numeric)) +
  theme_few(base_size=12) +
  geom_jitter(data = unique_labels, width=.1, height=.1, alpha=.1, aes(y=num_labels)) +
  geom_pointrange(data = unique_by_age_by_category, aes(ymin=ci_lower, ymax=ci_upper)) +
  geom_smooth(data = unique_by_age_by_category,  span=10, method='lm', color='grey') +
  scale_color_viridis_c() +
  facet_wrap(~category, nrow=2) +
  theme(legend.position='none', aspect.ratio=1) +
  ylim(0,8) +
  ylab('Number of unique labels') +
  xlab('Age of drawer')

ggsave(here::here('data/figures_out/Appendix_A8_unique_parts_by_age.pdf'), width=8, units='in')

write_csv(unique_by_age_by_category, file = here::here('data/figures_csvs/Appendix_A8_unique_parts_by_age.csv'))

       
```


## Inferential: Does the number of parts change with age?
```{r include=TRUE}
### Look at age effect in lmer add bobqa optimizer to make converge (still can't get random slopes, sad)

labels_by_age <- lmer(data=unique_labels, (num_labels) ~ scale(age_numeric) + (1|category),  control=lmerControl(optimizer="bobyqa",
                                 optCtrl=list(maxfun=2e5)))

labels_by_age_out = round(summary(labels_by_age)$coefficients,3)

round(confint(labels_by_age, method='Wald'),3)
```

## Text: describing unique parts x age
Consistent with this idea, we found that drawings produced by older children generally contained more unique semantic parts than drawings by younger children  (B=`r labels_by_age_out[2,1]`, SE = `r labels_by_age_out[2,2]`, df = `r labels_by_age_out[2,3]`, P<.001, CI=[.314, .476])



```{r}
img_per_category_age <- d_agree_labels %>%
  group_by(category, age_numeric) %>%
  summarize(count_images = length(unique(filename)))  %>%
  mutate(category = factor(category, levels = c("bird", "fish", "rabbit","dog" ,"sheep","tiger" ,"camel" , "bear", "cup","lamp", "hat","bottle", "car", "boat","airplane","train")))

top_parts_emphasis <- part_emphasis_all %>%
  group_by(category, roi_labelName) %>%
  filter(!roi_labelName %in% c('body','head')) %>%
  summarize(avg_emphasis = mean(emphasis), count_drawings = n()) %>%
  group_by(category) %>%
  slice_max(n=5, order_by=count_drawings)

part_emphasis_top_4_plot <- part_emphasis_all %>%
  right_join(top_parts_emphasis %>% select(category, roi_labelName)) %>%
  group_by(category, age_numeric, roi_labelName) %>%
  dplyr::summarize(avg_emphasis = mean(emphasis), count_drawings_with_part = n()) %>%
  left_join(img_per_category_age) %>%
  mutate(prop_drawings = count_drawings_with_part / count_images)

```

## Figure: Changes in part inclusion/emphasis by age
```{r}
ggplot(data = part_emphasis_top_4_plot, aes(x=age_numeric, y=prop_drawings, col=roi_labelName)) +
  geom_point(aes(size = avg_emphasis, alpha=.2)) +
  scale_size(range = c(0,4)) +
  theme_few(base_size=7, base_family = "Helvetica") +
  geom_line(stat='smooth',alpha=.2, span=10, aes(group=roi_labelName)) +
  ylab('Proportion drawings with object part') +
  xlab('Age of drawer') +
  theme(legend.position = 'none', aspect.ratio = 1)  +
  ylim(0,1) +
  facet_wrap(~category, nrow=2, scales = 'free_x') +
  ggrepel::geom_label_repel(data =part_emphasis_top_4_plot %>% filter(age_numeric==8), aes(label = roi_labelName), max.overlaps=10, size=1.8, label.size=.05, box.padding=0)

ggsave(here::here('data/figures_out/Figure4_part_emphasis_by_age.pdf'),units = 'in', height = 2.5, width = 7.09)

write_csv(part_emphasis_top_4_plot, file = here::here('data/figures_csvs/Figure4_part_emphasis_by_age.csv'))

```

## Appendix Figure: Changes in object part structure with VGG 48-way classification
```{r}
part_emphasis_dnn_recognition <- part_emphasis_all %>%
  mutate(sketch_path = as.factor(str_split_fixed(filename,'.png',2)[,1])) %>%
  left_join(d)  %>%
  filter(!is.na(correct_or_not))

part_emphasis_clip_recognition <- part_emphasis_all %>%
  mutate(sketch_path = as.factor(str_split_fixed(filename,'.png',2)[,1])) %>%
  left_join(d_clip)  %>%
  filter(!is.na(correct_or_not))
```

```{r}
# some drawings were marked invalid from the set by prolific workers but not by manual exclusions, meaning that they weren't included in the 37K set.

# 
# what <- part_emphasis_all %>%
#   mutate(sketch_path = as.factor(str_split_fixed(filename,'.png',2)[,1])) %>%
#   left_join(d_clip)  %>%
#   filter(is.na(correct_or_not)) %>%
#   ungroup() %>%
#   distinct(filename, category, age_numeric) %>%
#   group_by(category, age_numeric) %>%
#   summarize(num_excluded = n())
```

```{r}
# plus, a bunch of drawings were ecxluded from dnn classifications for balancing...annoying

# what <- part_emphasis_all %>%  
#   mutate(sketch_path = as.factor(str_split_fixed(filename,'.png',2)[,1])) %>%
#   left_join(d)  %>%
#   filter(is.na(correct_or_not)) %>%
#   ungroup() %>%
#   distinct(filename, category, age_numeric) %>%
#   group_by(category, age_numeric) %>%
#   summarize(num_excluded = n())

```

```{r}
part_emphasis_recognition_binned <-  part_emphasis_dnn_recognition %>%
  ungroup() %>%
  group_by(category) %>% # construct bins WITHIN categories
  mutate(model_correct_bin = ntile(log_odds,4))

part_emphasis_clip_recognition_binned <-  part_emphasis_clip_recognition %>%
  ungroup() %>%
  group_by(category) %>% # construct bins WITHIN categories
  mutate(model_correct_bin = ntile(log_odds,4))
```


```{r}
binned_count_recognition_by_sketch <- part_emphasis_recognition_binned %>%
  group_by(category, model_correct_bin) %>%
  summarize(count_images_bin = length(unique(filename)))

```

```{r}
binned_count_clip_recognition_by_sketch <- part_emphasis_clip_recognition_binned %>%
  group_by(category, model_correct_bin) %>%
  summarize(count_images_bin = length(unique(filename)))

```

```{r}
# get emphasized parts without recognition
top_parts_emphasis <- part_emphasis_all %>%
  group_by(category, roi_labelName) %>%
  filter(!roi_labelName %in% c('body','head')) %>%
  summarize(avg_emphasis = mean(emphasis), count_drawings = n()) %>%
  group_by(category) %>%
  slice_max(n=5, order_by=count_drawings)

```

```{r}
part_emphasis_top_4_by_model_bin <- part_emphasis_recognition_binned %>%
  right_join(top_parts_emphasis %>% select(category, roi_labelName)) %>%
  group_by(category, model_correct_bin, roi_labelName) %>%
  dplyr::summarize(avg_emphasis = mean(emphasis), count_drawings_with_part = length(unique(filename))) %>%
  left_join(binned_count_recognition_by_sketch) %>%
  mutate(prop_drawings = count_drawings_with_part / count_images_bin)
```

```{r}
part_emphasis_top_4_by_clip_model_bin <- part_emphasis_clip_recognition_binned %>%
  right_join(top_parts_emphasis %>% select(category, roi_labelName)) %>%
  group_by(category, model_correct_bin, roi_labelName) %>%
  dplyr::summarize(avg_emphasis = mean(emphasis), count_drawings_with_part = length(unique(filename))) %>%
  left_join(binned_count_clip_recognition_by_sketch) %>%
  mutate(prop_drawings = count_drawings_with_part / count_images_bin)
```

```{r}
# part_names = unique(part_emphasis_top_4_by_model_bin$roi_labelName)

part_emphasis_top_4_by_model_bin <- part_emphasis_top_4_by_model_bin %>%
  mutate(category = factor(category, levels = c("bird", "fish", "rabbit","dog" ,"sheep","tiger" ,"camel" , "bear", "cup","lamp", "hat","bottle", "car", "boat","airplane","train")))

part_emphasis_top_4_by_clip_model_bin <- part_emphasis_top_4_by_clip_model_bin %>%
  mutate(category = factor(category, levels = c("bird", "fish", "rabbit","dog" ,"sheep","tiger" ,"camel" , "bear", "cup","lamp", "hat","bottle", "car", "boat","airplane","train")))
```

```{r}
ggplot(data = part_emphasis_top_4_by_model_bin, aes(x=model_correct_bin, y=prop_drawings, col=roi_labelName)) +
  geom_point(aes(size = avg_emphasis, alpha=.2)) +
  facet_wrap(~category, nrow=1) +
  theme_few(base_size=8) +
  geom_line(stat='smooth',alpha=.2, span=10) +
  ylab('Proportion drawings with object part') +
  xlab('Binned classifier evidence (VGG-19 classifications)') +
  theme(legend.position = 'none', aspect.ratio = 1)  +
  ylim(0,1) +
  facet_wrap(~category, nrow=2, scales = 'free_x') +
  scale_x_continuous(
    limits=c(.5,4.5),
    breaks=c(1,4),
    labels=c(' Low ',' High ')
        ) +
  ggrepel::geom_label_repel(data =part_emphasis_top_4_by_model_bin %>% filter(model_correct_bin==1), aes(label = roi_labelName), max.overlaps=10, size=2, box.padding=.1, label.padding=.1)

ggsave(here::here('data/figures_out/Appendix_A9_part_emphasis_by_model_bin.pdf'),units = 'in', height = 3, width = 8)

write_csv(part_emphasis_top_4_by_model_bin, file = here::here('data/figures_csvs/Appendix_A9_part_emphasis_by_model_bin.csv'))

```


## Not included in Appendix, parts by CLIP model here for completeness.
```{r}
ggplot(data = part_emphasis_top_4_by_clip_model_bin, aes(x=model_correct_bin, y=prop_drawings, col=roi_labelName)) +
  geom_point(aes(size = avg_emphasis, alpha=.2)) +
  scale_size_area() +
  facet_wrap(~category, nrow=1) +
  theme_few(base_size=8) +
  geom_line(stat='smooth',alpha=.2, span=10) +
  ylab('Proportion drawings with object part') +
  xlab('Binned classifier evidence (CLIP classifications)') +
  theme(legend.position = 'none', aspect.ratio = 1)  +
  ylim(0,1) +
  facet_wrap(~category, nrow=2, scales = 'free_x') +
  scale_x_continuous(
    limits=c(.5,4.5),
    breaks=c(1,4),
    labels=c(' Low ',' High ')
        ) +
  ggrepel::geom_label_repel(data =part_emphasis_top_4_by_clip_model_bin %>% filter(model_correct_bin==1), aes(label = roi_labelName), max.overlaps=10, size=2, box.padding=.1, label.padding=.1)


```

# Session info
```{R}
pander(sessionInfo())
```
