==> analysis/step8_production_vs_recognition.Rmd <==
---
title: "Drawing production vs. recognition - exploratory analyses"
author: "Bria Long"
date: "6/20/2022, finalized 7/2023"
output:
  html_document:
    toc: true
    theme: united
---

# Libraries and data
```{r setup, include=FALSE}
library(tidyverse)
library(here)
library(assertthat)
library(ggthemes)
library(viridis)
library(lme4)
library(lmerTest)
library(langcog)
library(pander)
```

```{r include=FALSE}
knitr::opts_chunk$set(warning=FALSE, message=FALSE)
```


## Load metadata
```{r}
all_meta <- read_csv(file = here::here('data/drawings/stringent_cleaned_dataset_meta/all_object_metadata_cleaned.csv')) %>%
  as_tibble() %>%
  mutate(category = str_split_fixed(category,' ',2)[,2]) %>%
  mutate(category = str_replace(category,' ','.'))  # ice cream
```



## Load clip classification data
```{r}
load(file=here::here('data/clip_outputs/fullset/clip_by_image.RData'))
clip_by_image <-  clip_by_image %>%
  mutate(sketch_path = paste0(category, '_sketch_','age', kid_age, '_cdm_', session_id)) %>%
  rename(intended_category = category) %>%
  rename(producer_age_numeric = kid_age) %>%
  rename(clip_correct_or_not = correct_or_not) %>%
  rename(clip_prob = max_prob) %>%
  ungroup() %>%
  select(sketch_path, intended_category, producer_age_numeric,  clip_correct_or_not, clip_prob)

```

## Load recognition data (already joined with vgg classificationw)
```{r}
load(here::here('data/recognition_data/all_recognition_data_preprocessed.RData'))
recog_data = all_joined
rm(all_joined)
```

## Merge
```{r}
recog_data_by_both <- recog_data %>%
  mutate(producer_age_numeric = as.numeric(str_split_fixed(producer_age,'age',2)[,2])) %>%
  rename(vgg_correct_or_not = model_correct_or_not) %>%
  rename(vgg_log_odds = log_odds) %>%
  rename(vgg_distinct_index = distinct_index) %>%
  left_join(clip_by_image, by=c('sketch_path','intended_category', 'producer_age_numeric')) %>%
  filter(!is.na(clip_correct_or_not)) # don't have clip/vgg for every sketch -- some were filtered after inclusion in recoggames, 271 sketches

# mismatch <- recog_data_by_both %>%
  # filter(is.na(clip_correct_or_not))

  
```

# Choose which model

## Does clip or vgg do a better job of predicting kid recognition?
```{r}
by_image <- recog_data_by_both %>%
  group_by(sketch_path) %>%
  summarize(avg_kid_correct = mean(correct_or_not), clip_correct_or_not = mean(clip_correct_or_not), vgg_correct_or_not = mean(vgg_correct_or_not))


cor.test(by_image$avg_kid_correct, by_image$clip_correct_or_not)
cor.test(by_image$avg_kid_correct, by_image$vgg_correct_or_not)

diff = lm(data = by_image, avg_kid_correct ~ clip_correct_or_not*vgg_correct_or_not )
summary(diff)

```

```{r}
by_image_long <- by_image %>%
  pivot_longer(cols = clip_correct_or_not:vgg_correct_or_not, names_to='model', values_to = 'prop_recognized') %>%
  mutate(category = str_split_fixed(sketch_path,'_',2)[,1])

```

## Model vs. kid recognition by sketch
```{r}
ggplot(data = by_image_long, aes(x=prop_recognized, y=avg_kid_correct, color=model)) +
  geom_jitter(height=.1, alpha=.2) +
  geom_smooth(method='lm') + 
  facet_wrap(~category, nrow=4) +
  theme_few() +
  xlab('model recognition by sketch') +
  ylab('kid avg recognition by sketch')
```


## Create data structues

### By recognizer age

#### Compute recognititon x age x category  for top 30%
```{r}
recognizer_by_age_by_category <- recog_data_by_both %>%
  mutate(distinct_index_clip = ntile(clip_prob,10)) %>%
  filter(distinct_index_clip>7) %>%
  group_by(intended_category, recognizer_age_numeric) %>%
  summarize(avg_recognized = mean(correct_or_not), count = length(unique(sketch_path)))

# how well 3-year-olds recognized drawings of e.g., airplanes OF ALL AGES
```

## By producer age
###Compute averaged recognition x category/age by model
```{r}
clip_producer_age_by_category <- clip_by_image %>%
 filter(!sketch_path %in% unique(recog_data_by_both$sketch_path)) %>%
  filter(intended_category %in% unique(recog_data_by_both$intended_category)) %>%
  group_by(intended_category, producer_age_numeric) %>%
  summarize(avg_recognized = mean(clip_correct_or_not), count = length(unique(sketch_path)))  %>%
  mutate(mode = 'clip_recognition')
```


### Merge producer & recognizer age
```{r}
both_clip_long <- recognizer_by_age_by_category %>%
  rename(age_numeric = recognizer_age_numeric) %>%
  mutate(mode = 'kid_recognition') %>%
  full_join(clip_producer_age_by_category %>% rename(age_numeric = producer_age_numeric)) %>%
  filter(!is.na(avg_recognized))
  
both_clip <- both_clip_long %>%
  pivot_wider(values_from = c(count, avg_recognized), names_from = c('mode')) %>%
  mutate(category = fct_drop(intended_category)) %>%
  rename(kid = avg_recognized_kid_recognition, model = avg_recognized_clip_recognition ) %>%
  filter(!is.na(kid)) %>%
  mutate(drawings_total = count_kid_recognition + count_clip_recognition)


```

```{r}

## kid recognition: if a kid is 3-years-old, how many "airplane" drawings does he recognized on average
## kid production: if a kid is 3-year-old, how many "recognizable" airplanes drawings can he produce (on average)

```



# Plot it
## Figure 7: recognition by producton
```{R}
both_clip$age_factor = factor(both_clip$age_numeric, labels = c('3-year-olds','4-year-olds','5-year-olds','6-year-olds','7-year-olds','8-year-olds','9-year-olds','10-year-olds'))
```

```{r}
ggplot(both_clip, aes(x=kid, y=model, color=category, size=count_clip_recognition)) +
  geom_point(alpha=.4)  +
  theme_few(base_size=7, base_family = "Helvetica") +
  theme(aspect.ratio=1, legend.position = 'none')  +
  ylab('Drawing Production (Model recognition accuracy)') +
  xlab('Drawing Recognition (Children\'s recognition accuracy)')  +
  scale_y_continuous(breaks=c(0,.2,.4,.6,.8, 1), limits=c(0,1)) +
  scale_x_continuous(breaks=c(0,.2,.4,.6,.8, 1), limits=c(0,1))   +
  geom_smooth(method='lm',alpha=0, size=.5, aes(weight=count_clip_recognition), color='grey') + 
  # scale_color_viridis_c() +
  ggrepel::geom_label_repel(aes(label = category), size=2, max.overlaps = 20, force_pull=3,  label.padding = .1, box.padding = .1, label.size = .1,  alpha=.8, segment.color = NA)+
  facet_wrap(~age_factor, nrow=2)


ggsave(here::here('data/figures_out/Figure7_production_vs_recognition.pdf'), width=7.09, height=3.5, units='in')

ggsave(here::here('data/figures_out/Figure7_production_vs_recognition.pdf'), width=7.09, height=4, units='in')

# write_csv(both_clip, file = here::here('data/figures_csvs/Figure7_production_vs_recognition.csv'))

```

```{r}
sum(both_clip$count_kid_recognition)
```

```{r}
sum(both_clip$count_clip_recognition)
```

# Run basic correlation

```{r}
by_category = both_clip %>% 
  group_by(intended_category) %>%
  summarize(avg_kid = mean(kid), avg_model = mean(model))
```
```{r}
prod_recog_cor = cor.test(by_category$avg_kid, by_category$avg_model)

# $r$ = `r  round(prod_recog_cor$estimate,3)`, $t$(`r prod_recog_cor$parameter`)= `r prod_recog_cor$statistic`, $P$ < `r prod_recog_cor$p.value`, CI = [`r round(prod_recog_cor$conf.int[1],3`), `r round(prod_recog_cor$conf.int[2],3)`]
```

```{r}
pander(sessionInfo())
```
